pyproject.toml
[project]
name = "cognizant"
version = "0.1.0"
description = "An AI-powered application to simplify complex study materials using Gemini."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-core>=1.3.0",
    "langchain-google-genai>=4.2.2",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
